arc: Change default for -mepilogue-cfi / -mno-epilogue-cfi option.
The command line options -mepilogue-cfi and -mno-epilogue-cfi cause some
problems.  These options cause the FRAME_RELATED flag to be cleared on
all epilogue instructions.  As a result of this, no DWARF cfi tracking
is performed for epilogue instructions.

As a result of this lack of cfi tracking, the DWARF generated around the
epilogue of a function will frequently be wrong, and if GCC is built
with assert checking on, assertions will sometimes trigger during the
DWARF cfi generation, as these inconsistent states are spotted.

Consider the following example function:

    void
    some_function (void *arg)
    {
       if (arg == NULL)
         return;

       /* function body */

       return;
    }

This might be converted to something like the following, notice that the
function prologue and epilogue have been added, but the argument check
has been moved to before the prologue code:

    some_function:

       if (arg == NULL)
         goto function_return;

       /* function prologue */

       /* function body */

       /* function epilogue */

    function_return:

       return;

Now in this example, when generating the cfi information, a series of
traces are created, at the points where these traces join the cfi state
must be consistent no matter which path is taken to reach that point.
So in the example above, a trace is started at the entry point of
'some_function', this trace splits at the 'goto', with one branch
jumping directly to 'function_return' while the other branch passes
through the prologue, body, and epilogue, before finally re-joining the
original trace at 'function_return'.

The problem then is this, the original trace that reaches the 'goto' has
not yet seen any prologue instructions, as such the cfi state is the
default cfi state.  When the 'goto' is visited, this causes a new trace
point to be initialised at 'function_return', this trace point inherits
the cfi state from the 'goto', that is the default cfi state.

The trace then passes through the prologue, this modifies the cfi state,
storing registers to the stack, and updating the stack pointer.

When the trace passes through the epilogue, this reverses the effect of
the prologue, and should restore the trace to the default cfi state.
However, this only happens if the epilogue instructions are correctly
marked as being FRAME_RELATED.  When the epilogue instructions are not
marked as FRAME_RELATED, then the trace will arrive at the
'function_return' point with a non-default cfi state.  This triggers an
assertion from within GCC as the trace mechanism already knows that
'function_return' should have the default cfi state.

This patch is a first step to fixing this issue.  In this patch I
reverse the default state of the mechanism that clears the FRAME_RELATED
flag.  Previously the FRAME_RELATED flag was cleared by default, after
this commit the flag is not cleared by default, but can be cleared by
passing -mno-epilogue-cfi command line option.

I consider this to be a temporary measure, by changing the default bugs
will be exposed where the epilogue instructions are not correctly
marked, or the instructions themselves can't be handled (currently) by
the cfi tracking mechanism.  Later patches in this series will address
these issues, however, until all these issues have been resolved it
might be helpful to be able to revert to the old behaviour.

gcc/ChangeLog:

	* config/arc/arc.c (arc_expand_epilogue): Update option used to
	clear frame related flag.
	* config/arc/arc.opt: Update option used for clearing frame
	related flag from epilogue instructions.